rename huarunpeijin to reinforcement

+++ App/ThMEPStructure/Reinforcement/Model/ThLTypeEdgeComponent.cs
@@ -3,10 +3,10 @@
     public class ThLTypeEdgeComponent:ThEdgeComponent
     {
         // 请参见"2021-ZZ-技术-总部-JG-CKTJ.pdf Page70"
-        public int Bw { get; set; }
-        public int Hc1 { get; set; }
-        public int Bf { get; set; }
-        public int Hc2 { get; set; }
+        public double Bw { get; set; }
+        public double Hc1 { get; set; }
+        public double Bf { get; set; }
+        public double Hc2 { get; set; }
         /// <summary>
         /// 拉筋
         /// 对应标注图中的数字2
